+++ src/services/service_service.go
@@ -10,7 +10,7 @@ func GetServices() ([]*model.Service, *errors.AppError) {
 	return dao.GetServices()
 }
 
-func GetServiceById(serviceId int) (*model.Service, error) {
+func GetServiceById(serviceId int64) (*model.Service, error) {
 	service, err := dao.GetServicesById(serviceId)
 
 	if err != nil {
